processTransportData > transportMarker.js
